+++ textdisplay.cc
@@ -13,7 +13,7 @@ void TextDisplay::displayPlayer(int player) {
 	out << "Player " << (player + 1) << ":" << endl;
 	out << "Downloaded: "<< pi[player]->data << "D, "<< pi[player]->virus << "V" << endl;
 	int abilities = 0;
-	for (int i = 0; i < pi[player]->abilities.size(); ++i) {
+	for (size_t i = 0; i < pi[player]->abilities.size(); ++i) {
 		if(!pi[player]->abilityUsed[i]) ++abilities;
 	}
 	out << "Abilities: "<< abilities << endl;
